Mise à jour de la note globale utilisateur (#93)
* Update user global rating after a new rating was recorded

* User floats for user global ratings in test data

+++ src/main/java/fr/eql/teama/catalogue/dto/FullUserDto.java
@@ -20,7 +20,7 @@ public class FullUserDto {
     private Boolean premium;
     private Boolean admin;
     private String registrationNumber = "";
-    private Integer globalRating;
+    private Float globalRating;
     private String image = "";
     private String geolocation = "";
